feat: Changed repository to get primary key as Integer type, not Long

+++ src/main/java/com/practice/springbootdocker/repository/HospitalRepository.java
@@ -3,5 +3,5 @@ package com.practice.springbootdocker.repository;
 import com.practice.springbootdocker.domain.entity.Hospital;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HospitalRepository extends JpaRepository<Hospital, Long> {
+public interface HospitalRepository extends JpaRepository<Hospital, Integer> {
 }
